server/requirements.txt
python-dateutil>=2.5.3
flask>=0.11.1
numpy>=1.11.1
pandas>=0.22.0
pytz>=2016.6.1
requests>=2.10.0
pyzmq>=15.2.1
fastapi>=0.63.0
uvicorn>=0.13.4
fastapi-utils>=0.2.1